move fhttpd stats reporting to own file.
